refactor: apply clean architecture principles & prepare for rokon-qt
I did not want this to be a breaking change for packagers so I have kept rokon-gtk as the default compilation target. Additionally, the rokon binary no longer exists. It is now either rokon-gtk or rokon-qt. Considering the facts: The only package that had a opportunity to be used by bug testers was the RPM, it is the only one that actually obsoletes the `rokon` package.

The Flathub and Snapcraft packages need to be investigated more. It'd be a shame to not provide a version for both.